+++ crawlerwidget.h
@@ -47,7 +47,7 @@ class CrawlerWidget : public QSplitter
     // Loads the passed file and reports success.
     bool readFile( const QString& fileName, int topLine );
     // Get the size (in bytes) and number of lines in the current file.
-    void getFileInfo( int* fileSize, int* fileNbLine,
+    void getFileInfo( qint64* fileSize, int* fileNbLine,
        QDateTime* lastModified ) const;
     // Get the line number of the first line displayed.
     int getTopLine() const;
